requirements.txt
Django==5.1.4
anthropic==0.75.0
python-dotenv==1.2.1
djangorestframework==3.15.2
celery==5.4.0
redis==5.2.1
django-celery-results==2.5.1
django-celery-beat==2.7.0
